Merged main & fixes
forgot to merge main before pushing..
So I did that, then fixed conflicts.
Also gametypes feature passes the game type when room is created so i remove the middleware that deals with getting the game type because its no longer necessary. Fixed the createdRoom call at createRoom component, it sent wrong values to the redux reducer.
This is now up to date with main and fully working as per my checks..